Remove unimplemented Workflow action type and optimize module structure
### Removed Workflow Support:
- **ActionType::Workflow** variant and all related code
- Workflow pattern matching in all execution functions
- `create_workflow_rule` test helpers and `test_decide_action_workflow`
- Workflow support in `compile_action` function
- Updated example config to use SendKeys instead

### Optimized Module Structure:
- **Moved `has_data_to_process`** from `trigger/data_checker.rs` to `periodic.rs`
- **Deleted `data_checker.rs`** - function now private within PeriodicTaskManager
- Better cohesion: data checking logic now co-located with periodic task management

### Benefits:
- **Eliminated dead/unimplemented functionality** (Workflow TODOs)
- **Reduced module complexity** - removed unnecessary abstraction layer
- **Improved code locality** - related functions grouped together
- **Cleaner, more focused codebase** with better separation of concerns
- All tests passing (27 unit + 3 integration tests)
