+++ src/trigger/periodic.rs
@@ -1,19 +1,21 @@
+use anyhow::Result;
+use std::process::Command;
 use std::sync::Arc;
 use tokio::task::JoinHandle;
 use tokio::time::interval;
 
-use crate::agent;
-use crate::cli::{execute_periodic_entry, has_data_to_process};
+use crate::agent::AgentPool;
+use crate::agent::execution::execute_periodic_entry;
 use crate::config::entry::{CompiledEntry, TriggerType};
 
 /// Periodic task manager responsible for handling periodic entries
 pub struct PeriodicTaskManager {
     pub entries: Vec<CompiledEntry>,
-    pub agent_pool: Arc<agent::AgentPool>,
+    pub agent_pool: Arc<AgentPool>,
 }
 
 impl PeriodicTaskManager {
-    pub fn new(entries: Vec<CompiledEntry>, agent_pool: Arc<agent::AgentPool>) -> Self {
+    pub fn new(entries: Vec<CompiledEntry>, agent_pool: Arc<AgentPool>) -> Self {
         Self {
             entries,
             agent_pool,
@@ -90,5 +92,34 @@ impl PeriodicTaskManager {
                 }
             }
         })
+    }
+}
+
+/// Check if a periodic entry will produce data to process
+async fn has_data_to_process(entry: &CompiledEntry) -> Result<bool> {
+    // If there's no source command, we consider it as having data to process
+    if entry.source.is_none() {
+        return Ok(true);
+    }
+
+    if let Some(source) = &entry.source {
+        // Execute the source command
+        let output = Command::new("sh").arg("-c").arg(source).output()?;
+
+        if !output.status.success() {
+            return Ok(false);
+        }
+
+        let stdout = String::from_utf8_lossy(&output.stdout);
+        let lines: Vec<String> = stdout
+            .lines()
+            .filter(|line| !line.trim().is_empty())
+            .map(|s| s.to_string())
+            .collect();
+
+        // Return true if we have any lines to process
+        Ok(!lines.is_empty())
+    } else {
+        Ok(true)
     }
 }
